pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "giga_auto_qc"
description = "A package"
readme = "README.md"
requires-python = ">=3.8"
license = { text="MIT" }
authors = [
  { name="Hao-Ting Wang", email="htwangtw@gmail.com" },
]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
]
dependencies = [
  "nilearn",
  "pybids<0.16.0",
  "templateflow<23.0.0",
]
dynamic = ["version"]

[project.scripts]
giga_auto_qc = "giga_auto_qc.run:main"

[project.optional-dependencies]
dev = [
  "black",
  "pre-commit",
  "giga_auto_qc[test]",
]
test = [
  "pytest",
  "pytest-cov",
]
# Aliases
tests = ["giga_auto_qc[test]"]


[project.urls]
"Homepage" = "https://github.com/htwangtw/giga_auto_qc"

[tool.hatch.version]
source = "vcs"

[tool.hatch.build.hooks.vcs]
version-file = "giga_auto_qc/_version.py"

[tool.hatch.build.targets.sdist]
exclude = [".git_archival.txt"]

[tool.hatch.build.targets.wheel]
packages = ["giga_auto_qc"]

[tool.black]
target-version = ['py38']
exclude = "giga_auto_qc/_version.py"
line_length = 79
